Moved examples into subfolders
